+++ MUMENT/MUMENT/Sources/Components/MumentCardBySongTVC.swift
@@ -30,6 +30,10 @@ class MumentCardBySongTVC: UITableViewCell {
     func setData(_ cellData: MumentCardBySongModel){
         mumentCard.setData(cellData)
     }
+    
+    func setData(_ cellData: AllMumentsResponseModel.MumentList){
+        mumentCard.setData(cellData)
+    }
 }
 
 // MARK: - UI
